requirements.txt
Django==4.2.5
environs[django]==9.5.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
factory-boy==3.3.0
PyJWT==2.8.0
flake8==6.1.0
isort==5.12.0
psycopg2-binary==2.9.7